[+] distance Map is now int and not short

+++ include/ia/placement.h
@@ -11,7 +11,7 @@
 #include "tools.h"
 
 t_vector *player_drop(t_player_info *player, char **map);
-void move_to_target(t_player_info *player, short **, t_vector *pos);
+void move_to_target(t_player_info *player, int **, t_vector *pos);
 void placement(char **map, t_player_info *player);
 
 #endif //PSU_LEMIPC_2017_PLACEMENT_H
